requirements.txt
# jupyter
jupyterlab

# openai
openai
tiktoken

torchmetrics

hydra-core

# misc
python-dotenv
tenacity
seaborn

streamlit
streamlit-option-menu
streamlit-authenticator

krippendorff